Remove redundant C++ standard sets from CMakeFiles
These were introduced in 098fb77611e54a9829428c253c60b2697e77f590 - did
Leo not know that they were already there at the top of the file? This
does the same thing, except it only sets it for VVVVVV instead of
everything (so this wouldn't set it for the third-party dependencies).